refactor: removed repetitive logout function and added single on topnavbar
